requirements.txt
requests
lxml
Pillow
beautifulsoup4
